+++ Sibusten.Philomena.Client.Examples/DownloadSvgImages.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Sibusten.Philomena.Client.Extensions;
 using Sibusten.Philomena.Client.Images;
 using Sibusten.Philomena.Client.Options;
 
@@ -20,6 +21,16 @@ namespace Sibusten.Philomena.Client.Examples
             });
             IPhilomenaImageDownloader downloader = new PhilomenaImageDownloader(search);
             await downloader.DownloadAllToFilesAsync(image => $"ExampleDownloads/DownloadSvgImages/{image.Id}.{image.Format}");
+
+            // Alternatively, with fluent extensions
+            await client
+                .Search("original_format:svg", new ImageSearchOptions
+                {
+                    SvgMode = SvgMode.Both,
+                    MaxImages = 5
+                })
+                .GetDownloader()
+                .DownloadAllToFilesAsync(image => $"ExampleDownloads/DownloadSvgImages/{image.Id}.{image.Format}");
         }
     }
 }